Break SGObject into multiple mixins
